[IMP] logging
Add support for logging to Discord through webhooks. A new logger is available
to post messages to Discord without printing to the console nor a log file,
also allowing to pass in embed objects for better styling in the Discord client.

Several logs have been improved in term of both the information printed and the styling
in the console or in Discord.

Some logging levels have been changed to keep clarity in the console and Discord.