[IMP] collated balance display
Switch from fetching the current balance trhough the API
to using the cached balances while forcing a complete refresh
when displaying the current balance.

When displaying the balance, also show the change (in percentage)
since starting the process.

Improve visual display of the balance both in Discord and in the console

- Display one currency per line
- Compute alignment to avoid large blank spaces